json-ast/cs: extract AST types
